requirements.txt
boto3
botocore
numpy==2.2.0
opencv-python==4.10.0.84
PyYAML==6.0.2
